Reverter endringer av legacy pdl-klient
