+++ src/main/kotlin/no/nav/sosialhjelp/innsyn/session/SessionController.kt
@@ -1,17 +1,17 @@
 package no.nav.sosialhjelp.innsyn.session
 
 import kotlinx.coroutines.runBlocking
-import no.nav.security.token.support.core.api.ProtectedWithClaims
-import no.nav.sosialhjelp.innsyn.app.subjecthandler.SubjectHandlerUtils
+import kotlinx.coroutines.slf4j.MDCContext
+import kotlinx.coroutines.withContext
+import no.nav.sosialhjelp.innsyn.app.subjecthandler.SubjectHandlerUtils.getUserIdFromToken
+import no.nav.sosialhjelp.innsyn.digisossak.hendelser.RequestAttributesContext
 import no.nav.sosialhjelp.innsyn.pdl.PdlService
-import no.nav.sosialhjelp.innsyn.utils.IntegrationUtils.ACR_IDPORTEN_LOA_HIGH
-import no.nav.sosialhjelp.innsyn.utils.IntegrationUtils.ACR_LEVEL4
-import no.nav.sosialhjelp.innsyn.utils.IntegrationUtils.SELVBETJENING
+import no.nav.sosialhjelp.innsyn.app.protectionAnnotation.ProtectionSelvbetjeningHigh
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 
-@ProtectedWithClaims(issuer = SELVBETJENING, claimMap = [ACR_LEVEL4, ACR_IDPORTEN_LOA_HIGH], combineWithOr = true)
+@ProtectionSelvbetjeningHigh
 @RestController
 @RequestMapping("/api/v1/innsyn")
 class SessionController(
@@ -19,7 +19,18 @@ class SessionController(
 ) {
     @GetMapping("/session")
     fun getSessionInfo(): SessionInfo = runBlocking {
-        val pid = SubjectHandlerUtils.getUserIdFromToken()
-        SessionInfo(fornavn = pdlService.getFornavnByPid(pid), harAdressebeskyttelse = pdlService.getAdressebeskyttelseByPid(pid))
+        withContext(MDCContext() + RequestAttributesContext()) {
+            with(getUserIdFromToken()) {
+                SessionInfo(
+                    fornavn = pdlService.getFornavnByPid(this),
+                    harAdressebeskyttelse = pdlService.getAdressebeskyttelseByPid(this)
+                )
+            }
+        }
     }
+
+    data class SessionInfo(
+        val fornavn: String,
+        val harAdressebeskyttelse: Boolean,
+    )
 }
